src/data/data_loader.py
import numpy as np
from src.models.intensity_reconstruction import IntensityReconstructor

class DataLoader:
    def __init__(self, params):
        self.params = params
        self.reconstructor = IntensityReconstructor(params)

    def generate_synthetic_data(self):
        """
        Generate synthetic data based on the current parameters.
        """
        x, y, t, X, Y, T = self.reconstructor.define_space_time_grid()
        phase = self.reconstructor.calculate_phase_modulation(X, Y, T)
        intensity = self.reconstructor.generate_interference_intensity(phase)
        measurement_mask, measurement_indices = self.reconstructor.define_measurement_region(X, Y)
        y_measurements = self.reconstructor.get_intensity_measurements(intensity, measurement_indices)
        
        return {
            'X': X,
            'Y': Y,
            'T': T,
            'intensity': intensity,
            'measurement_mask': measurement_mask,
            'measurement_indices': measurement_indices,
            'y_measurements': y_measurements
        }

    def load_experimental_data(self, file_path):
        """
        Load experimental data from a file.
        This is a placeholder method and should be implemented based on your specific data format.
        """
        # TODO: Implement this method based on your experimental data format
        raise NotImplementedError("Loading experimental data is not yet implemented.")

    def get_data(self, use_synthetic=True, file_path=None):
        """
        Get either synthetic or experimental data.
        """
        if use_synthetic:
            return self.generate_synthetic_data()
        else:
            if file_path is None:
                raise ValueError("File path must be provided for experimental data.")
            return self.load_experimental_data(file_path)